feat(status, date filter): core deer status, date filter nemlee
